+++ resources/release-notes.md
@@ -4,6 +4,7 @@
   ```
   settings().MAG.Toolbox.Sound
   ```
+- (All) Fix issue with `datetime` range slider showing the wrong time range for HK plots
 
 ## Software